requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
asttokens=2.2.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.2=pypi_0
configargparse=1.5.3=pypi_0
contourpy=1.0.6=pypi_0
cryptography=38.0.1=py38h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
dash=2.7.1=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
debugpy=1.6.4=pypi_0
decorator=5.1.1=pypi_0
entrypoints=0.4=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.16.2=pypi_0
ffmpeg=4.3=hf484d3e_0
flask=2.2.2=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2022.11.0=pypi_0
fvcore=0.1.5.post20210915=py38
giflib=5.2.1=h7b6447c_0
gin-config=0.5.0=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.15.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.51.1=pypi_0
idna=3.4=py38h06a4308_0
imageio=2.22.4=pypi_0
importlib-metadata=5.1.0=pypi_0
importlib-resources=5.10.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=py38
ipykernel=6.19.2=pypi_0
ipython=8.7.0=pypi_0
ipywidgets=8.0.3=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
jsonschema=4.17.3=pypi_0
jupyter-client=7.4.8=pypi_0
jupyter-core=5.1.0=pypi_0
jupyterlab-widgets=3.0.4=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=8.2.0=hdf63c60_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libopenblas=0.3.2=h9ac9557_1
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lightning-utilities=0.4.2=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
minkowskiengine=0.5.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=6.0.3=pypi_0
nbformat=5.5.0=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=pypi_0
nettle=3.7.3=hbbd107a_1
numpy=1.23.4=py38h14f4228_0
numpy-base=1.23.4=py38h31eccc5_0
oauthlib=3.2.2=pypi_0
open3d=0.16.0=pypi_0
openblas-devel=0.3.2=0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=22.0=pypi_0
pandas=1.5.2=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=py38hace64e9_1
pip=22.3.1=py38h06a4308_0
piqa=1.2.2=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=2.6.0=pypi_0
plenoxel=0.0.1.dev0+sphtexcub.lincolor.fast=pypi_0
plotly=5.11.0=pypi_0
portalocker=2.6.0=py38h578d9bd_1
prompt-toolkit=3.0.36=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.13.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pyquaternion=0.9.9=pypi_0
pyrsistent=0.19.2=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.15=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=1.12.1=py3.8_cuda11.3_cudnn8.3.2_0
pytorch-lightning=1.8.5.post0=pypi_0
pytorch-mutex=1.0=cuda
pytorch3d=0.7.1=pypi_0
pytz=2022.6=pypi_0
pyyaml=6.0=py38h0a891b7_4
pyzmq=24.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=py38h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.0=pypi_0
scipy=1.9.3=pypi_0
setuptools=65.5.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.0=h5082296_0
stack-data=0.6.2=pypi_0
tabulate=0.9.0=pyhd8ed1ab_1
tenacity=8.1.0=pypi_0
tensorboard=2.11.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5.1=pypi_0
termcolor=2.1.1=pyhd8ed1ab_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.1=py38_cu113
torchmetrics=0.11.0=pypi_0
torchvision=0.13.1=py38_cu113
tornado=6.2=pypi_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.7.1=pypi_0
typing_extensions=4.4.0=py38h06a4308_0
urllib3=1.26.13=py38h06a4308_0
wcwidth=0.2.5=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.4=pypi_0
xz=5.2.8=h5eee18b_0
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h7f98852_2
yarl=1.8.2=pypi_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
